feat: StyledProgressBarMolecule
